0.5.8-1 verify fields(store)

+++ src/main/java/dev/qwett/webappspring/service/impl/StoreServiceImpl.java
@@ -25,39 +25,67 @@ public class StoreServiceImpl implements StoreService {
     }
 
     public Store findById(int id) {
+        if (id < 0) {
+            logger.warn("invalid id , {} < 0", id);
+            return null;
+        }
         logger.info("Find Store with id = {}", id);
         return storeRepository.findById(id).orElse(null);
     }
 
     public Store addStore(Store store) {
         logger.info("Add store...");
-        if (storeRepository.findById(store.getIdStore()).orElse(null) == null) {
-            logger.info("Store with address = {} , was added", store.getAddress());
-            return storeRepository.save(store);
+        if (store.getAddress() != null && !store.getAddress().trim().isEmpty()) {
+            if (storeRepository.findById(store.getIdStore()).orElse(null) == null) {
+                store = storeRepository.save(store);
+                logger.info("Store with address = {} , was added", store.getAddress());
+                return store;
+            }
+            logger.warn("Adding error occurred, Store address = {}, this store is already in DB", store.getAddress());
+            return null;
         }
-        logger.warn("Adding error occurred, Store address = {}", store.getAddress());
+        logger.warn("Adding error occurred, Field address is empty");
         return null;
     }
 
     public Store updateStore(int id, Store store) {
-        store.setIdStore(id);
         logger.info("Updating Store with id = {}", id);
-        if (storeRepository.findById(store.getIdStore()) != null) {
-            logger.info("Store id = {}, with Address = {},  was updated", id, store.getAddress());
-            return storeRepository.save(store);
-        }
-        logger.warn("Updating error occurred, Store id = {}, Address = {}", id, store.getAddress());
+        if (id < 0) {
+            logger.warn("invalid id , {} < 0", id);
+        } else if (store.getAddress() != null && !store.getAddress().trim().isEmpty()) {
+            store.setIdStore(id);
+            if (storeRepository.findById(store.getIdStore()) != null) {
+                store = storeRepository.save(store);
+                logger.info("Store id = {}, with Address = {},  was updated", id, store.getAddress());
+                return store;
+            }
+            logger.warn("Updating error occurred, Store id = {}, Address = {}, no field found (ID)", id, store.getAddress());
+            return null;
+        } else
+            logger.warn("Field address is empty, update failed");
         return null;
     }
 
     @Override
     public List<Store> findByAddress(String address) {
-        logger.info("Search for entries, where store address = {}", address);
-        return storeRepository.findByAddress(address);
+        if (address != null && !address.trim().isEmpty()) {
+            logger.info("Search for entries, where store address = {}", address);
+            return storeRepository.findByAddress(address);
+        }
+        logger.warn("Field address is empty");
+        return null;
     }
 
     public void delete(int id) {
-        storeRepository.deleteById(id);
-        logger.info("Store id = {}, was deleted", id);
+        if (id < 0)
+            logger.warn("invalid id , {} < 0", id);
+        else {
+            long repoCount = storeRepository.count();
+            storeRepository.deleteById(id);
+            if (repoCount > storeRepository.count())
+                logger.info("Store id = {}, was deleted", id);
+            else
+                logger.warn("Store id = {} is not deleted", id);
+        }
     }
 }
